Replace process.versions tests by browserify's package.json browser field. Fixes #329
