+++ lib/BIP32.js
@@ -4,7 +4,7 @@ var coinUtil = imports.coinUtil || require('../util');
 var Key = imports.Key || require('./Key');
 var Point = imports.Point || require('./Point');
 var SecureRandom = imports.SecureRandom || require('./SecureRandom');
-var bignum = imports.bignum || require('./Bignum');
+var bignum = imports.bignum || require('bignum');
 var networks = require('../networks');
 
 var secp256k1_n = new bignum('FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141', 16);
