+++ lib/Base58.js
@@ -1,5 +1,5 @@
 var crypto = require('crypto');
-var bignum = require('./Bignum');
+var bignum = require('bignum');
 
 var globalBuffer = new Buffer(1024);
 var zerobuf = new Buffer(0);
