--- lib/Bignum.js
@@ -1,5 +1,0 @@
-if (process.versions) {
-  module.exports = require('bignum');
-  return;
-}
-module.exports = require('./browser/Bignum');
