+++ lib/Block.js
@@ -3,7 +3,7 @@ var imports            = require('soop').imports();
 var util              = imports.util || require('../util');
 var Debug1            = imports.Debug1 || function() {};
 var Script            = imports.Script || require('./Script');
-var Bignum            = imports.Bignum || require('./Bignum');
+var Bignum            = imports.Bignum || require('bignum');
 var Binary            = imports.Binary || require('binary');
 var Step              = imports.Step || require('step');
 var buffertools       = imports.buffertools || require('buffertools');
