+++ lib/Curve.js
@@ -1,6 +1,6 @@
 "use strict";
 var imports = require('soop');
-var bignum = imports.bignum || require('./Bignum');
+var bignum = imports.bignum || require('bignum');
 var Point = imports.Point || require('./Point');
 
 var n = bignum.fromBuffer(new Buffer("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141", 'hex'), {size: 32});
