+++ lib/Electrum.js
@@ -2,7 +2,7 @@ var Key = require('./Key'),
   Point = require('./Point'),
   twoSha256 = require('../util').twoSha256,
   buffertools = require('buffertools'),
-  bignum = require('./Bignum');
+  bignum = require('bignum');
 
 /**
  * Pre-BIP32 Electrum public key derivation (electrum <2.0)
