+++ lib/Key.js
@@ -1,5 +1,1 @@
-if (process.versions) {
-  module.exports = require('./node/Key');
-  return;
-}
-module.exports = require('./browser/Key');
+module.exports = require('bindings')('KeyModule').Key;
