+++ lib/ScriptInterpreter.js
@@ -4,7 +4,7 @@ var log = imports.log || require('../util/log');
 var util = imports.util || require('../util');
 var Opcode = imports.Opcode || require('./Opcode');
 var buffertools = imports.buffertools || require('buffertools');
-var bignum = imports.bignum || require('./Bignum');
+var bignum = imports.bignum || require('bignum');
 var Util = imports.Util || require('../util');
 var Script = require('./Script');
 var Key = require('./Key');
