+++ lib/SecureRandom.js
@@ -1,5 +1,10 @@
-if (process.versions) {
-  module.exports = require('./node/SecureRandom');
-  return;
+var imports = require('soop');
+var crypto = imports.crypto || require('crypto');
+
+var SecureRandom = require('./common/SecureRandom');
+
+SecureRandom.getRandomBuffer = function(size) {
+  return crypto.randomBytes(size);
 }
-module.exports = require('./browser/SecureRandom');
+
+module.exports = require('soop')(SecureRandom);
