+++ lib/Transaction.js
@@ -5,7 +5,7 @@ var Address = imports.Address || require('./Address');
 var Script = imports.Script || require('./Script');
 var ScriptInterpreter = imports.ScriptInterpreter || require('./ScriptInterpreter');
 var util = imports.util || require('../util');
-var bignum = imports.bignum || require('./Bignum');
+var bignum = imports.bignum || require('bignum');
 var Put = imports.Put || require('bufferput');
 var Parser = imports.Parser || require('../util/BinaryParser');
 var Step = imports.Step || require('step');
