+++ lib/TransactionBuilder.js
@@ -54,7 +54,7 @@ var imports = require('soop').imports();
 var Address = imports.Address || require('./Address');
 var Script = imports.Script || require('./Script');
 var util = imports.util || require('../util');
-var bignum = imports.bignum || require('./Bignum');
+var bignum = imports.bignum || require('bignum');
 var buffertools = imports.buffertools || require('buffertools');
 var networks = imports.networks || require('../networks');
 var WalletKey = imports.WalletKey || require('./WalletKey');
