+++ lib/browser/Key.js
@@ -1,7 +1,7 @@
 var ECKey = require('../../browser/vendor-bundle.js').ECKey;
 var SecureRandom = require('../SecureRandom');
 var Curve = require('../Curve');
-var bignum = require('../Bignum');
+var bignum = require('bignum');
 
 var Key = function() {
   this._pub = null;
