+++ lib/browser/Point.js
@@ -2,7 +2,7 @@
 
 var imports = require('soop').imports();
 var Key = imports.Key || require('./Key'); 
-var bignum = imports.bignum || require('../Bignum');
+var bignum = imports.bignum || require('bignum');
 var assert = require('assert');
 var ECPointFp = require('../../browser/vendor-bundle.js').ECPointFp;
 var ECFieldElementFp = require('../../browser/vendor-bundle.js').ECFieldElementFp;
