--- lib/node/Key.js
@@ -1,3 +1,0 @@
-var Key = require('bindings')('KeyModule').Key;
-
-module.exports = Key;
